fix: make some fixes in client and add loading spinner for calendar

+++ src/client/src/components/change-password/index.js
@@ -1,36 +1,33 @@
 import React, { useState, useEffect, useRef } from "react";
-import "../change-password/styles.css";
+import "../change-password/styles.scss";
 import Button from "components/shared/Button";
 import api from "services/api";
 import ModalWindow from "components/shared/ModalWindow";
 import { Formik } from "formik";
 import Input from "components/shared/Input";
 
-
 export const ChangePassword = (tool) => {
-
   const [isOpen, setisOpen] = useState(true);
 
   function ChangePassw(values) {
     if (values.newPass1 == values.newPass2) {
-      api.post(`/auth/change-password`, {
-        "oldPassword": values.oldPass,
-        "newPassword": values.newPass,
-        "confirmNewPassword": values.confirmNewPassword,
-      })
-        .then((response) => {
-          setisOpen(false)
-          localStorage.setItem("isFirstLogin","false")
-
+      api
+        .post(`/auth/change-password`, {
+          oldPassword: values.oldPass,
+          newPassword: values.newPass,
+          confirmNewPassword: values.confirmNewPassword,
         })
-        .catch((error) => { });
+        .then((response) => {
+          setisOpen(false);
+          localStorage.setItem("isFirstLogin", "false");
+        })
+        .catch((error) => {});
     }
   }
   const initialValuesPass = {
     oldPass: "",
     newPass: "",
     confirmNewPassword: "",
-
   };
 
   return (
@@ -60,7 +57,6 @@ export const ChangePassword = (tool) => {
           }) => (
             <form onSubmit={handleSubmit}>
               <div>
-
                 <div className="row">
                   <Input
                     onChange={(e) => {
@@ -71,7 +67,7 @@ export const ChangePassword = (tool) => {
                     name="oldPass"
                     placeholder="Старый пароль"
                     onBlur={handleBlur}
-                    autocomplete="off"
+                    autoComplete="off"
                   />
                 </div>
                 <div className="row">
@@ -84,7 +80,7 @@ export const ChangePassword = (tool) => {
                     name="newPass"
                     placeholder="Новый пароль"
                     onBlur={handleBlur}
-                    autocomplete="off"
+                    autoComplete="off"
                   />
                 </div>
                 <div className="row">
@@ -97,15 +93,19 @@ export const ChangePassword = (tool) => {
                     name="confirmNewPassword"
                     placeholder="Новый пароль"
                     onBlur={handleBlur}
-                    autocomplete="off"
+                    autoComplete="off"
                   />
-                </div>  
+                </div>
                 <div className="row">
-                  <Button onClick={() => { ChangePassw(values) }} >
+                  <Button
+                    onClick={() => {
+                      ChangePassw(values);
+                    }}
+                  >
                     Сохранить
                   </Button>
-                </div> 
-              </div> 
+                </div>
+              </div>
             </form>
           )}
         </Formik>
